# internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for SchoolMaterialStatus.
const (
	SchoolMaterialStatusActive  SchoolMaterialStatus = "active"
	SchoolMaterialStatusArchive SchoolMaterialStatus = "archive"
)

// Defines values for SchoolMaterialPayloadStatus.
const (
	SchoolMaterialPayloadStatusActive  SchoolMaterialPayloadStatus = "active"
	SchoolMaterialPayloadStatusArchive SchoolMaterialPayloadStatus = "archive"
)

// Defines values for MaterialType.
const (
	Article      MaterialType = "article"
	Presentation MaterialType = "presentation"
	Video        MaterialType = "video"
)

// SchoolMaterial defines model for SchoolMaterial.
type SchoolMaterial struct {
	// Content Текстовое описание материала
	Content string `json:"content"`

	// CreatedAt Дата создания
	CreatedAt time.Time          `json:"created_at"`
	Id        openapi_types.UUID `json:"id"`

	// MaterialType Тип материала
	MaterialType MaterialType `json:"material_type"`

	// Name Краткое название материала
	Name string `json:"name"`

	// Status Статус публикации
	Status SchoolMaterialStatus `json:"status"`

	// UpdatedAt Дата изменения
	UpdatedAt time.Time `json:"updated_at"`
}

// SchoolMaterialStatus Статус публикации
type SchoolMaterialStatus string

// SchoolMaterialArray defines model for SchoolMaterialArray.
type SchoolMaterialArray struct {
	Materials []SchoolMaterial `json:"materials"`
}

// SchoolMaterialPayload defines model for SchoolMaterialPayload.
type SchoolMaterialPayload struct {
	Content string `json:"content"`

	// MaterialType Тип материала
	MaterialType MaterialType                `json:"material_type"`
	Name         string                      `json:"name"`
	Status       SchoolMaterialPayloadStatus `json:"status"`
}

// SchoolMaterialPayloadStatus defines model for SchoolMaterialPayload.Status.
type SchoolMaterialPayloadStatus string

// MaterialType Тип материала
type MaterialType string

// SchoolMaterialArrayResponseOK defines model for SchoolMaterialArrayResponseOK.
type SchoolMaterialArrayResponseOK = SchoolMaterialArray

// SchoolMaterialResponseOK defines model for SchoolMaterialResponseOK.
type SchoolMaterialResponseOK = SchoolMaterial

// UUIDResponseOK defines model for UUIDResponseOK.
type UUIDResponseOK struct {
	Uuid *openapi_types.UUID `json:"uuid,omitempty"`
}

// NewSchoolMaterial defines model for NewSchoolMaterial.
type NewSchoolMaterial = SchoolMaterialPayload

// UpdateSchoolMaterial defines model for UpdateSchoolMaterial.
type UpdateSchoolMaterial = SchoolMaterialPayload

// GetSchoolMaterialArrayParams defines parameters for GetSchoolMaterialArray.
type GetSchoolMaterialArrayParams struct {
	MaterialType  *MaterialType `form:"material_type,omitempty" json:"material_type,omitempty"`
	FromCreatedAt *time.Time    `form:"from_created_at,omitempty" json:"from_created_at,omitempty"`
	ToCreatedAt   *time.Time    `form:"to_created_at,omitempty" json:"to_created_at,omitempty"`
	Page          *int          `form:"page,omitempty" json:"page,omitempty"`
	PageSize      *int          `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// CreateSchoolMaterialJSONRequestBody defines body for CreateSchoolMaterial for application/json ContentType.
type CreateSchoolMaterialJSONRequestBody = SchoolMaterialPayload

// UpdateSchoolMaterialJSONRequestBody defines body for UpdateSchoolMaterial for application/json ContentType.
type UpdateSchoolMaterialJSONRequestBody = SchoolMaterialPayload